refactor: SectionService 분리
