+++ src/main/java/nextstep/subway/applicaion/LineService.java
@@ -1,5 +1,6 @@
 package nextstep.subway.applicaion;
 
+import lombok.RequiredArgsConstructor;
 import nextstep.subway.applicaion.dto.LineRequest;
 import nextstep.subway.applicaion.dto.LineResponse;
 import nextstep.subway.applicaion.dto.SectionRequest;
@@ -16,18 +17,13 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
+@Transactional
+@RequiredArgsConstructor
 @Service
-@Transactional(readOnly = true)
 public class LineService {
-    private LineRepository lineRepository;
-    private StationService stationService;
+    private final LineRepository lineRepository;
+    private final StationService stationService;
 
-    public LineService(LineRepository lineRepository, StationService stationService) {
-        this.lineRepository = lineRepository;
-        this.stationService = stationService;
-    }
-
-    @Transactional
     public LineResponse saveLine(LineRequest request) {
         Line line = lineRepository.save(new Line(request.getName(), request.getColor()));
         if (request.getUpStationId() != null && request.getDownStationId() != null && request.getDistance() != 0) {
@@ -38,34 +34,25 @@ public class LineService {
         return createLineResponse(line);
     }
 
+    @Transactional(readOnly = true)
     public List<LineResponse> showLines() {
         return lineRepository.findAll().stream()
                              .map(this::createLineResponse)
                              .collect(Collectors.toList());
     }
 
+    @Transactional(readOnly = true)
     public LineResponse findById(Long id) {
         return createLineResponse(lineRepository.findById(id).orElseThrow(IllegalArgumentException::new));
     }
 
-    @Transactional
     public void updateLine(Long id, LineRequest lineRequest) {
         Line line = lineRepository.findById(id).orElseThrow(IllegalArgumentException::new);
         line.update(lineRequest.getName(), lineRequest.getColor());
     }
 
-    @Transactional
     public void deleteLine(Long id) {
         lineRepository.deleteById(id);
-    }
-
-    @Transactional
-    public void addSection(Long lineId, SectionRequest sectionRequest) {
-        Station upStation = stationService.findById(sectionRequest.getUpStationId());
-        Station downStation = stationService.findById(sectionRequest.getDownStationId());
-        Line line = lineRepository.findById(lineId).orElseThrow(IllegalArgumentException::new);
-
-        line.addSection(new Section(line, upStation, downStation, sectionRequest.getDistance()));
     }
 
     private LineResponse createLineResponse(Line line) {
@@ -88,13 +75,5 @@ public class LineService {
         return stations.stream()
                        .map(it -> stationService.createStationResponse(it))
                        .collect(Collectors.toList());
-    }
-
-    @Transactional
-    public void deleteSection(Long lineId, Long stationId) {
-        Sections sections = lineRepository.findById(lineId)
-                                          .orElseThrow(IllegalArgumentException::new)
-                                          .getSections();
-        sections.removeSection(stationId);
     }
 }
