+++ src/test/java/nextstep/subway/unit/LineServiceMockTest.java
@@ -6,7 +6,6 @@ import nextstep.subway.applicaion.dto.LineRequest;
 import nextstep.subway.applicaion.dto.SectionRequest;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.LineRepository;
-import nextstep.subway.domain.Section;
 import nextstep.subway.domain.Station;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -31,25 +30,6 @@ public class LineServiceMockTest {
     @InjectMocks
     private LineService lineService;
 
-    @Test
-    void addSection() {
-        // given
-        final Station 기흥역 = new Station(1L, "기흥역");
-        final Station 신갈역 = new Station(2L, "신갈역");
-        given(stationService.findById(1L)).willReturn(기흥역);
-        given(stationService.findById(2L)).willReturn(신갈역);
-
-        final Line line = new Line(3L, "분당선", "yellow");
-        given(lineRepository.findById(3L)).willReturn(Optional.of(line));
-
-        // when
-        lineService.addSection(line.getId(), new SectionRequest(기흥역.getId(), 신갈역.getId(), 10));
-
-        // then
-        assertThat(line.getSections().size()).isEqualTo(1);
-        assertThat(line.getSections().getStationNames()).containsExactlyInAnyOrder("기흥역", "신갈역");
-    }
-
     @ParameterizedTest
     @CsvSource(value = {"에버라인:red:에버라인:red", "에버라인::에버라인:yellow", ":red:분당선:red"}, delimiter = ':')
     void updateLine(String lineName, String color, String expectLineName, String expectColor) {
@@ -61,31 +41,9 @@ public class LineServiceMockTest {
         lineService.updateLine(line.getId(), LineRequest.builder().color(expectColor).name(expectLineName).build());
 
         // then
-        assertThat(line.getName()).isEqualTo(expectLineName);
-        assertThat(line.getColor()).isEqualTo(expectColor);
-    }
-
-    @Test
-    void deleteSection() {
-        // given
-        final Station 기흥역 = new Station(1L, "기흥역");
-        final Station 신갈역 = new Station(2L, "신갈역");
-        final Station 정자역 = new Station(3L, "정자역");
-
-        final Line line = new Line(4L, "분당선", "yellow");
-
-        line.addSection(기흥역, 신갈역, 10);
-        line.addSection(신갈역, 정자역, 10);
-
-        given(lineRepository.findById(line.getId())).willReturn(Optional.of(line));
-
-        // when
-        lineService.deleteSection(line.getId(), 정자역.getId());
-
-        // then
         assertAll(
-            () -> assertThat(line.getSections().size()).isEqualTo(1),
-            () -> assertThat(line.getSections().getStationNames()).containsExactlyInAnyOrder("기흥역", "신갈역")
+            () -> assertThat(line.getName()).isEqualTo(expectLineName),
+            () -> assertThat(line.getColor()).isEqualTo(expectColor)
                  );
     }
 }
